Refactor : moved all java files to src/main/java
