--- src/BuyableLot.java
@@ -1,11 +1,0 @@
-import com.modeliosoft.modelio.javadesigner.annotations.objid;
-
-@objid ("0e16b588-e350-4c47-9c73-56cecb8edc50")
-public abstract class BuyableLot {
-
-    ///Methods
-    @objid ("d0013750-53cb-41d7-92bf-490d9b451930")
-    public abstract void OnOwnershipChange() {
-    }
-
-}
