--- src/FreeState.java
@@ -1,5 +1,0 @@
-import com.modeliosoft.modelio.javadesigner.annotations.objid;
-
-@objid ("f73b84a3-8ae1-4e23-b153-3b919a010ade")
-public class FreeState extends PropertyState {
-}
