--- src/FullyConstructedState.java
@@ -1,5 +1,0 @@
-import com.modeliosoft.modelio.javadesigner.annotations.objid;
-
-@objid ("4231760f-6bbf-4e7e-99d1-cec29d20c47f")
-public class FullyConstructedState extends PropertyState {
-}
