--- src/PropertyState.java
@@ -1,17 +1,0 @@
-import com.modeliosoft.modelio.javadesigner.annotations.objid;
-
-@objid ("8b632616-cc58-492c-886b-c3e940ff50c4")
-public abstract class PropertyState {
-    @objid ("674c3e3f-8150-4f9b-9b52-e0126b5bac33")
-    public void Build() {
-    }
-
-    @objid ("8b951008-0f83-4c85-84d2-ea1f5fbd9b47")
-    public void SellBuilding() {
-    }
-
-    @objid ("9c1aefd1-500d-4a7e-9394-5abfb44d892b")
-    public void ApplyOnStop(Player player) {
-    }
-
-}
