--- src/SellableState.java
@@ -1,5 +1,0 @@
-import com.modeliosoft.modelio.javadesigner.annotations.objid;
-
-@objid ("aff905e8-fcd9-49b9-9b04-7e39fb1c2eda")
-public abstract class SellableState extends PropertyState {
-}
