+++ src/main/java/ConstructedState.java
@@ -1,0 +1,5 @@
+import com.modeliosoft.modelio.javadesigner.annotations.objid;
+
+@objid ("e723f9d3-b899-4fa6-ac45-8e609b597b87")
+public class ConstructedState extends PropertyState {
+}
