+++ src/main/java/ConstructibleState.java
@@ -1,0 +1,5 @@
+import com.modeliosoft.modelio.javadesigner.annotations.objid;
+
+@objid ("cc7f0ed3-f688-4d5c-b342-07ab16338f86")
+public class ConstructibleState extends SellableState {
+}
