+++ src/main/java/OwnedState.java
@@ -1,0 +1,5 @@
+import com.modeliosoft.modelio.javadesigner.annotations.objid;
+
+@objid ("03502d46-4d79-470f-9514-40f1dcec835f")
+public class OwnedState extends SellableState {
+}
