+++ src/main/java/ParkTile.java
@@ -1,0 +1,24 @@
+import com.modeliosoft.modelio.javadesigner.annotations.objid;
+
+@objid ("847b2d1e-0d8a-4e85-b99b-4d3f6e5f7c23")
+public class ParkTile extends Tile {
+
+
+    ///Constructors
+    public ParkTile()
+    {
+        this.tileName = "Free Park Tile";
+    }
+
+    ///Methods
+    @objid ("0c9cb401-eb8c-44aa-88cd-3adce8e72120")
+    public void ApplyOnStop(Player player) {
+        //DO NOTHING
+    }
+
+    @objid ("65c318cb-bcc4-405f-8df5-0d95d498bc5d")
+    public void ApplyOnPassBy(Player player) {
+        //DO NOTHING
+    }
+
+}
