+++ src/main/java/PrisonTile.java
@@ -1,0 +1,23 @@
+import com.modeliosoft.modelio.javadesigner.annotations.objid;
+
+@objid ("b390ca95-12e1-4275-b80a-95ed5385359f")
+public class PrisonTile extends Tile {
+
+    ///Constructors
+    public PrisonTile()
+    {
+        this.tileName = "Prison Tile";
+    }
+
+    ///Methods
+    @objid ("2789d778-888a-445d-b973-969a43a3b25a")
+    public void ApplyOnStop(Player player) {
+        //DO NOTHING
+    }
+
+    @objid ("ade0b2a6-5f9b-4acc-b8e9-a497ba6f33c2")
+    public void ApplyOnPassBy(Player player) {
+        //DO NOTHING
+    }
+
+}
